+++ sys/arm/amlogic/aml8726/aml8726_pinctrl.c
@@ -210,7 +210,7 @@ aml8726_pinctrl_configure_pins(device_t dev, phandle_t cfgxref)
 
 	if (f->name == NULL) {
 		device_printf(dev, "unknown function attribute %.*s in FDT\n",
-		    len, function_name);
+		    (int) len, function_name);
 		OF_prop_free(function_name);
 		return (ENXIO);
 	}
@@ -234,7 +234,7 @@ aml8726_pinctrl_configure_pins(device_t dev, phandle_t cfgxref)
 		else {
 			device_printf(dev,
 			    "unknown pull attribute %.*s in FDT\n",
-			    len, pull);
+			    (int) len, pull);
 			OF_prop_free(pull);
 			return (ENXIO);
 		}
